tests/wpt/tests/css/css-align/parsing/align-tracks-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Grid Level 3: parsing align-tracks with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3/#tracks-alignment">
<meta name="assert" content="align-tracks supports the full grammar 'normal | <baseline-position> | <content-distribution> | <overflow-position>? <content-position>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("align-tracks", "normal");

// <baseline-position> = [ first | last ]? baseline
test_valid_value("align-tracks", "baseline");
test_valid_value("align-tracks", "first baseline", "baseline");
test_valid_value("align-tracks", "last baseline");

// <content-distribution> = space-between | space-around | space-evenly | stretch
test_valid_value("align-tracks", "space-between");
test_valid_value("align-tracks", "space-around");
test_valid_value("align-tracks", "space-evenly");
test_valid_value("align-tracks", "stretch");

// <overflow-position>? <content-position>
// <overflow-position> = unsafe | safe
// <content-position> = center | start | end | flex-start | flex-end
test_valid_value("align-tracks", "center");
test_valid_value("align-tracks", "start");
test_valid_value("align-tracks", "end");
test_valid_value("align-tracks", "flex-start");
test_valid_value("align-tracks", "flex-end");
test_valid_value("align-tracks", "unsafe end");
test_valid_value("align-tracks", "safe flex-start");

test_valid_value("align-tracks", "flex-start, last baseline");
test_valid_value("align-tracks", "normal, normal", "normal");
test_valid_value("align-tracks", "start, flex-end, flex-end, flex-end, flex-end", "start, flex-end");

</script>
</body>
</html>
